Assertion refactor (#148)
Changes assertion checking to be done by a separate transform, instead of interleaved with the main compiler. This also removes the assertion state from the compiler, which was redundant in any case with the final design.

Adds unit tests to check assertion behavior, both in the failed and missing cases.