Compiler forking (#170)
To support design space exploration, this allows a partial compilation
to be forked (repeatedly) so the results can then be continued with
additional refinements like block and parameter refinements, without
needing to recompile everything from scratch.

On scanning through the 3.3v regulator on the SMU, this reduces the time
per point from ~4s to ~0.7s. The savings are likely to significantly
vary by design and search configuration.

Infrastructural changes:
- Adds initFrom to ConstProp, DependencyGraph that allows cloning data
from one to another. These just check that the target is empty, and
copies over the state variables from another object.
- Support freezing parameters on ConstProp, which stops propagation into
(and past) the target parameter. This hasn't been tested with equality
(since the update behavior may need to be different), so that's just
require-checked right now.
- Add internal non-updating API for ConstProp to allow forced params to
be bulk-set before the first update kicks in and prevents the forced
params.
- Change forcedParams to store the target expr, so the setForcedValue
doesn't need to immediately update, and it can support forced values
after a prior assign is inserted (but before the assign sets the value).
- Add cloned for all wir objects. For immutable objects, this returns
self; for mutable objects this recursively clones.
- Adds concatenation for PartialCompile and filter-partition for
Refinements
- Compiler has different rules for init and fork
- And of course unit tests.
- Also some opportunistic cleanups.