Deterministic netlist, commit unit test netlists (#182)
Makes netlists deterministic with net nodes in roughly the order
connections are written. Adds the generated netlists from the examples
to git so changes are visible through the git changes interface (and
hopefully make it possible to catch unintentional functional changes).

Determinism is from a combination of proto ordering (in a prior PR) and
changing a set here to a list. Netlisting unit tests updated to have
ordered nets.

To preserve (roughly) written order, edge traversal (for building the
netlist) is changed to BFS from DFS.

Also adds better indenting to nets in generated netlists.