Add PCB tstamp / path upgrader (#190)
Older PCB designs used a component tstamp / path format that was the
string path which is no longer supported in KiCad 6 which requires the
hex format. So boards designed in KiCad 5 don't preserve components when
importing a new netlist. This adds a script that takes a PCB and
netlist, and updates the PCB tstamps / paths.

(note: while tstamps is what's indicated in the KiCad GUI, loading a new
netlist actually seems to link by the path field)

This also updates the CanAdapter PCB. The design is not changed, but
this updates the tstamp / path and does a KiCad 5 -> 6 format
conversion. Actually updating the netlist may change the design because
libraries have changed since then.