Separate SWO/TDI debug header pinning (#193)
Prior, SWO was part of SWD even for micros that do not support it. Now,
SWD is defined as only power, reset, SWCLK, SWDIO, and SWO and TDI (pins
available on the 10-pin SWD header) can be separately connected. Also
defines a mixin-style class for IoControllers with a SWD header.

By default, SWO and TDI on the header is not connected, though it can be
specified with a refinement. For now, all the examples are updated to
keep the previous netlist (except for blinky, which now has SWO
disconnected), though this could be changed later, eg connect a
bidirectional UART instead of the SWO pin.

Additionally, changes the compiler so a boundary array port can't be
directly exported from an internal allocated port. Examples in practice
don't use this so no code changes are required, but the semantics aren't
defined so it should fail noisily. Practically, because a mapping of
requested ports -> actual defined elements isn't available, it's not
possible to resolve that mapping, so for now the user most specify that
elementwise, as is done to support the new SWO/TDI behavior (we need to
connect both the externally requested GPIO and optional SWO/TDI pins to
the internal GPIO - intuitively we should be able to do
`connect(self.gpio, self.ic.gpio.request_vector())`, but that doesn't
work for reasons above so it's unrolled in a generator.