Directed assigns in connects, delay assign until declaration seen (#203)
Core compiler and semantics simplification. This changes the prior
bidirectional equality between parameters in connects to towards-link
directed assigns - which allows the complex bidirectional propagation
system to be removed.

This also delays assignments until the target parameter's declaration is
seen (for parameters only, things like .CONNECTED_LINK aren't gated).
This allows class-based value refinements to work, because the
assignment in the design tree is seen and processed before the block
class is known and the refinement can set the forced value.

Minor fixes to libraries to make this work, mostly removing where
parameters are over-assigned. The only sticking point is for buck
converters, where in concept the power path should generate the output
voltage, but we need the output voltage to solve for downstream current
which allows the power path to generate. The solution here is to use a
forced-voltage pseudo-block which presents the voltage before the power
path generates.