Capacitor voltage-rating derating (#204)
General rule of thumb for ceramic capacitors is to specify a voltage
rating of twice the operating voltage. This adds that functionality.

This doesn't change the current hacktastic capacitance derating (where
if the capacitor is operating under high voltage, more nameplate
capacitance is requested to offset capacitance derating).

In practice this doesn't have seemed to changed any of the part
selections, perhaps because the default capacitors are already pretty
high voltage.

Also adds a note about why generators only take arg-params.