Clean up categories (#212)
Add categories where there are none and fix some ones that exist, so
that the blocks library panel looks clean and usable. Some general
cleanup too.

Other fixes:
- In general hide a bunch of inner stuff behind an `InternalBlock`
category superclass
- Allow adapters to be created on boundary ports
- When subclassing a `non_library` superclass, the superclass's
superclasses are pulled in to the proto (and won't throw an error for
things like refinements if it's missing a link).
  - Add unit test
- Replaces a bunch of `abstract_block` with `non_library`. The general
rule is that `abstract_block` should be able to be instantiated, while
`non_library` is for utility classes that provide implementation
functionality but should not be instantiated. There is a boundary case
where some leftover non-instantiable `abstract_block` classes are left
for categorization purposes in the library viewer.
  - This includes a bunch of the Table* and *StandardPinning blocks
- Create abstract CanEsdDiode, with one implementation (Pesd1can)
- Create abstract UsbDevice and Host connectors, and have the existing
blocks implement these
- Move old solar-car blocks into test_high_switch - out of the main
libraries that are maintained and indexed by the IDE