Custom schematic import blackbox handler (#240)
Generalized infrastructure for power users to define a custom blackbox
handler, eg to support additional fields. It's kinda ugly actually, but
for now it gets the job done in an architecturally elegant way.

Adds a custom JLC blackbox hander that understands the 'JLCPCB Part #'
field that appears to be used by some plugins.