Add packaging tests, support blocks defined in __main__ (#252)
Add local packaging run to workflow tests, with an abbreviated
(blinky-only) unittest and top-level test.

Resolves #251, where blocks defined in `__main__` cause an error when
the HDL server tries to resolve those blocks, because the HDL server is
a separate Python instance and thinks itself is `__main__` (instead of
the original file). The fix is to detect when a library's module is
`__main__` and effectively guess the correct original module name by
using the filename. Might be a better way to do it, but this works, and
is largely what a websearch turned out for finding the original module
name for `__main__`.

Bumps the package version since this is a significant bugfix.