Don't apply capacitor derating by default (#253)
Capacitor capacitance derating is one of those unintuitive things and
(probably?) isn't done in applications where an exact capacitance is
calculated (e.g. most capacitors on schematics seem to be specified as
nominal; in contrast to power systems where there are equations that
produce an exact capacitance to meet some spec). This adds an
`exact_capacitance` parameter for Capacitor (and DecouplingCapacitor),
which indicates whether an exact capacitance was specified (and should
do derating). It is off by default, and only used in power converters in
the standard library.

Removes the old `derate_capacitance` flag, which really wasn't used, is
a bit technical, and didn't fit into libraries (was designed as a
top-level refinement control).

Some of the netlists update as a result.