SWD Probe v2 + fix refinement blocks with additional ports (#254)
A spin of the SWD debug probe using modern infrastructure. Major board
changes:
- Move the SWD source header into the board file. Technically it's
reusable, but it's a bit of a mess.
- Add a SWD source header for the TC2030 as a 6-pin 2.54mm dual-row
edge-inline header
- Add 5v protection Zener diode, add USB ESD diode
- Replace RGBs with single-color LEDs - while RGBs are nice we never did
anything with them in firmware
- Specify LED colors: white for USB connectivity (custom function),
yellow for programming activity, green for target power
- Add pull-up resistor on target LDO enable, so it should work with
stock firmware
- Add target voltage sense divider
- Use 0402 components
- Name pin assignments using port names instead of pin names, which
should be more stable
- Single-placement-side autorouted board, as an incremental test
platform. It's not great, a future spin will do a double-placement-side
more compact board
- Not yet fabricated, will be part of a larger panel

Library improvements:
- Add infrastructural TagConnect PassiveConnector (for 6, 8, 10
positions), replacing the prior scheme which they would be a direct
footprint
- Has explicitly legged and non-legged variants, with a default
refinement to the legged variant
- Add "official" TC2030 pinning for SWD target, in addition to the
custom TC2050 pinning
- The base SWD target block has TDI which is kind of wrong, but needs
better infrastructural support for mixins (#104) to improve
- Unify the Ap2204k blocks into one that has an optional enable pin (as
a generator that ties it high if not externally connected)
- Add a 2.54mm dual-row edge-inline header, with a corresponding
footprint. 6-pin only.
- These resulted in non-functional netlist changes.

Infrastructural bugfiixes:
- When a refined block has additional ports, those are now marked
not-connected and no requested. (prior, these variables were not given a
value, and generators relying on them would fail)
- Update the precompiled compiler jar