Automatically materialize USB and CAN mixins for compatibility (#283)
Legacy code had CAN and USB on all IoControllers, but mixins changed
that (because USB and CAN are not universal peripherals) which was a
breaking change. This automatically materializes the USB and CAN mixins
when requested using legacy-style code, by overloading
`__getattr__(...)`. This should also correctly encode the requirement
for mixins as part of the block type.

Likely examples and tutorials will continue using the old-style
(non-mixin) USB syntax for simplicity.