Expand connections in sub-expressions and preserve original expressions (#285)
Infrastructure for a larger feature, this preserves the original connect
expressions instead of lowering them in-place (as is prior behavior),
instead adding the lowered expressions to a expanded field in the
connect expression.

This keeps the connect implementation the same and should not
functionally change anything. Though with the new expanded field, it's
apparent that the previous implementation was pretty messy and made
heavy use of mutable state. Eventually, that should be fixed / totally
refactored and overhauled, but this isn't the time and place for it.
That might better track when a connect can be fully expanded, instead of
incrementally mutating the connect and hoping that it all works out at
the end.

In particular, there's multiple overlapping meanings for intermediate
expanded values in connectArray. Empty expanded could mean an empty
connect, or it hasn't been expanded. Single expanded could mean it
expanded into one connection, or a substitution was done on the
collective array connect, and still needs to be expanded into multiple
individual connects. In the future, these should be explicitly tracked
in the compiler.

This changes the IR in a way that breaks compatibility.

Also some simplification of internal compiler structures.