Better netlist component path pruning (#297)
Prior, component pruning ignored only adapters and bridges when
determining when an only child could be collapsed directly into its
parent's path. Now, this ignores any block with no children and that is
not a footprint - a heuristic for any adapter-like block which won't
have internal footprints.

Also moves the LCD outlines from the Device to the containing
application circuit, with the device being the connector only.
Potentially a bit weird from a hierarchy perspective, but better
supports the select-hierarchy-sheet operation in KiCad layout editor.