Zener-clamped high-side switch (#304)
Migrate high-side switch to a schematic-defined block, and add a
variation that has a zener clamp on the gate, for when the input voltage
is greater than a reasonable gate voltage. For now, the user has to
manually opt-in by specifying a clamp_voltage to HighSideSwitch.

Update DRV8313 to ground the NC pins and unused comparator output, per
the datasheet example layout.