Restructure into subpackages (#352)
Instead of a bunch of top-level packages, moves things like `edg_core`
into `edg.core`. Also renames some subpackages for simplicity:
`electronics_abstract_parts` -> `abstract_parts`, `electronics_lib` ->
`parts`. Should not break compatibility for those using `from edg import
*`, but will break for those importing internals.

This structure should allow this to be used as a submodule and is
overall cleaner. The compiler will detect if PolymorphicBlocks exists in
the current path, and if so, uses the qualified path.

`examples` is still independent and is not part of the public API or
package structure.