Compile progress (#353)
Adds an optional progressFn to Compiler, which is regularly called and
updated with the overall compilation progress.

Each block has some progress value, defined as the total compilation
progress (0-1) completed when the block and its subtree are fully
compiled. The top block's progress value is 1, and each block's progress
is the parent's progress divided by the immediate number of siblings
plus 1 (the extra 1 is for the parent itself - updated when the block is
compiled).

In practice this isn't linear, since units of work get progressively
smaller, but it's probably better than nothing.