Allow unnamed elements in error messages (#361)
Prior, exceptions that use _name_of that had an unknown element would
stack-trace with the _name_of. This could happen, for instance, during
an invalid implicit connect before the Block is assigned a name. This
adds an optional argument to _name_of to auto-generate unknown names for
elements.

The generate unknown names behavior isn't default for cases where the
name must be valid (as opposed to only being helpfully informative).