Nondirectioned Ground refactor (#364)
This introduces a dedicated non-directioned Ground type, with an
optional GroundReference which can specify the voltage level. Ground can
optionally specify a voltage tolerance.

Prior, Ground was actually just a VoltageSink in disguise, which had a
few problems stemming from the VoltageSink/VoltageSource directionality
that this fixes:
- Avoids needing explicit VoltageSource merges when connecting the
ground of multiple power supplies, eg USB + battery
- Avoids a confusing gnd and gnd_src for devices that can source or sink
power
- Eliminates unused current_draw parameter

This refactors a lot of internal libraries. Notable structural changes:
- BLDC driver current-sense resistors moved into the block, generating
if the sense pins are used.
- Ditto with the soldering iron connector
- Gate driver switch node is now a Ground
- Unify gnd and gnd_src for blocks that have both
- DecouplingCapacitor defined in terms of gnd/voltage instead of pos/neg

This change is breaking and requires some changes to top-level designs:
- Voltage-source merges for grounds for multiple power sources must be
removed
- GroundSource should be replaced with Ground (remapped with a
deprecation)
- VoltageTestPoint must be replaced with GroundTestPoint, now that the
types are different
- In general, ground-side current sense resistors are more difficult
now, the move so far is to move them into the associated block

Library cleanup to support offset grounds is needed in the future, to
make sure all devices use VoltageSink.from_gnd to use a ground-relative
voltage instead of assuming ground is 0v.

Also some internal cleanup.

Notes for the future:
- Some things that might need further thought
- Is there a clean unified way to get the voltage of a port accounting
for connected-ness and directionality? Currently, we have
`GroundLink._voltage_range` and `VoltageLink,_voltage_range`, though
code that doesn't need to be as robust uses port.link().voltage