New library parts and example boards: SX1262 LoRa, PN7160 NFC, FLIR Lepton thermal camera sensor (#384)
Adds these libraries:
- Bidirectional level shifter using two FETs , requires a directionality
hint
- A lot of RF library components / generators, including L networks, and
a L network overlaid with a second-harmonic notch
  - Unit tests for these verified against online calculators
- DiscreteRfWarning mixin, override-able assertion warning for discrete
RF blocks
- TLV757P high current SOT-23 LDOs
- PN7160 NFC controller with analog frontend generator. Experimental,
but tested working (though not necessary optimally tuned)
  - Unit tests for these verified against reference documents
- SX1262 LoRa transceiver, based on the reference analog frontend
architecture but with generators. The component values don't line up
exactly but are in the right ballpark.

Example boards:
- Standing desk controller - flipped RX/TX as sent to fab, corrected in
libraries and generator but not layout. Layout is still good enough
since the level shifters are bidirectional. Level shifter resistor for
the Neopixels also need to be 1k not 4.7k.
- SX1262 LoRa transceiver / PN7160 NFC omnibus RF test board - missing
the SX1262 BUSY pin connection, missing the PN7160 IRQ pin connection as
sent to fab, corrected in libraries and generator but not layout.
- FLIR Lepton test board - works with no modifications.

Changes to existing libraries:
- Add gnd to Antenna interface
- Add RF connectors as antennas, including Amphenol901143 SMA-F
connector
- Add SOT-323 to BJTs, FETs
- Un-comment DCR criteria for inductors
- Fixes for new JlcParts parts selector libraries
- Optional reset pull for OV2640, only generated if reset not externally
connected
- Expansion of SG8101 crystals to also 3.2x2.5mm devices
- Add support for using strapping pins of ESP32-C3-WROOM module
- More common default threshold for APX803s POR generator

TODO - future work
- Separate Shutdown-able and Resettable