Wrapper Footprint Blocks (#385)
Wrapper footprint blocks replace its internal contents with a footprint
during netlisting. That is, it can have internal blocks for electronics
model purposes, but when generating a board only its footprint shows up.

This is implemented as board scopes in the netlister. Potential
pre-feature for #367, daughterboard / multi-board assembly support.

Adds two examples, a new Pololu A4988 stepper driver board, and refactor
the existing VL53L0x connector to use this (instead of the weird
superclass stuff). In the future, perhaps the microcontrollers can also
be refactored to use this, though might need changes to the pin
remapping layer.

There are in two styles, with different trade-offs:
- The A4988 is a custom block that only instantiates the internal IC
using it for modeling. It doesn't create a bunch of ghost supporting
passives (if it has extended / instantiated the application circuit),
but duplicates a lot of code. Note, had this extended the application
circuit, the internal resistor would create a basic-part error.
- The VL53L0x extends the application circuit, which makes it a valid
refinement target and duplicates no code, but generates phantom passives
that don't get netlisted. Inheritance is stylistically ugly here.
- Unclear what the long-term solution should be.

Other changes:
- Empty nets are pruned during generation (this changes a bunch of
netlists)
- Refdesing is scope aware
- Example cleanups for the VL53L0x change