Add shrink-multiply op to backend (#402)
Add the tolerance-shrinking multiply operation to the backend, so it can
be done without generators which were constraining in terms of parameter
dataflow and syntactic boilerplate.

Renames the previous cancel_multiply (which was a bit too
implementation-focused) to shrink_multiply (which is more use-case
focused). Also better documentation in terms of target tolerance and
contributing tolerance.

Netlists should be the same, but there are ordering changes from changed
order of parts.

Resolves #393