Refactor core calculation out of buck / boost converters (#407)
This makes the core component calculations more unit-testable and
unifies the buck-boost converter (as well as allowing other
variations-on-a-theme converters like forward-buck-reverse-boost as with
some NVDC battery chargers with OTG with less code duplication)

Changes the boost inductor calculation to avoid double-counting output
tolerance. The inductance formula appears monotonic.

Core infrastructural changes:
- Refactor NumLikeExpr types (IntExpr, FloatExpr, RangeExpr) to
centrally define castable-types and properly return NotImplemented on a
type it cannot handle. This cleanly handles promotion of FloatLike /
IntLike types to Range types.
- Add hull and intersect to Range

Follow-on PRs will add coupled current/inductance support (#405) and the
MP2722 converter which this is building towards.