Standardize power path current draw conventions (#411)
Standardize conventions defined in the Buck/BoostConverterPowerPath:
switch on buck is an VoltageSink that draws the average inductor
current, and switch on the boost is a VoltageSource that provides the
max average output current. Refactors the BuckBoostConverterPowerPath to
use this convention. This isn't the cleanest, but this allows these
parameters to propagate to the connected device (eg, buck converter
controller IC's power-in pin) through the connection instead of needing
to be plumbed explicitly. This makes the custom discrete converter a bit
tougher, but probably a good trade-off.

Additional fixes:
- current limit on the boost power path, accounts for the boost ratio
properly.
- account for efficiency factor on current draw for buck and boost
converters
  - updates examples to tolerate this
- plumb rds_on through the custom converters
- model switch current limits in custom converters (generally not the
limiting factor, but still good practice)
  - add ForcedVoltageCurrentLimit as a library block
- unit tests that run the buck and boost converter power path and check
generated parameters

Adds the actual inductor current peak for all converter power paths.
Makes the discrete custom buck converter use this, instead of
calculating it independently.

Resolves #406 and #307