Add PowerSO8 / DFN5x6 fets (#421)
Add PowerSO8 / DFN5x6 FET pinnings using the generic footprint and add
package names for JLC parts tables. The naming is a mess, but the all
should be compatible packages.

Note that there are so few of these in the JLC table that they do not
get picked in practice in any of the examples.