Refactor Neopixel to add decoupling caps, use consistent pwr port naming (#428)
The old .vdd port naming is kept as an alias, eventually needs a proper
deprecation warning on access. Cleans up Neopixel docstrings to be less
redundant.

Also changes svg-pcb generation to use the first matching footprint if
there are multiple in a subtree. This unbreaks examples from changes.
Convention is the first block (recursively) is the anchor block, eg the
main chip.

Updates generated netlists.

Resolves #424