Prefer lower e-series values for resistors and capacitors (#430)
The series is added to a sort in the resistor and capacitor table. The
idea is to prefer more common parts where they fit. Inductors not
modified since they are much more nonstandard.

Removes the "E1" series, since it is nonstandard and promotes capacitors
jumping from 4.7uF to 10uF. Also removes the 24+192 series and the like,
those have a nonstandard key. The higher series are not used in the
examples, and if desired some similar function can be added in the
future.

In practice (from the netlist diffs), this picks better resistors. These
may be more concerning:
- This changes RF and crystal caps. TBD if that is problematic, but they
are still within their stated tolerance, and in any case it was not
selecting intelligently before at all.
- The SMU example needs to be updated to up the min clamp current,
otherwise the resistance picked is too high for the sink.

Fixes some OLED capacitor definitions to tighten the range. The new rule
is only the 2.2uF caps are allowed to extend, but not the 4.7uF which is
considered more standard.

Resolves #85