Enable strict typing with generics (#438)
Adds default type parameters to generics to allow most code to work
while the "default" generic class is a base class of sorts. This makes a
lot of type parameters covariant.

The exception is Vector, which allows mutable elements, and remains
invariant. BaseVector (non-parameterized) serves as the base class.

This also properly parameterizes the StandardFootprint type, so the
block type in the lambdas are checked. HasStandardFootprint also now
requires a StandardFootprint type-parameterzed on Self. To avoid
circular dependencies, the _STANDARD_FOOTPRINT can be a lambda.

This uses typing_extensions for TypeVar defaults and Self for earlier
version of Python.

Also removes OrderedDict and ABCMeta and the IR type parameter for
Block.